Make 'bayes_auto_learn_on_error' configurable
With bayes_auto_learn_on_error off, autolearning will be performed even
if bayes classifier already agrees with the new classification (i.e.
yielded BAYES_00 for what we are now trying to teach it as ham, or
yielded BAYES_99 for spam). This is a traditional setting, the default
was chosen to retain backward compatibility.

With bayes_auto_learn_on_error turned on, autolearning will be performed
only when a bayes classifier had a different opinion from what the
autolearner is now trying to teach it (i.e. it made an error in
judgement). This strategy may or may not produce better future
classifications, but usually works very well, while also preventing
unnecessary overlearning and slows down database growth.